% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_API.R
\name{estimate_gamma_bic}
\alias{estimate_gamma_bic}
\title{Estimate the PCF gamma parameter using BIC.}
\usage{
estimate_gamma_bic(x)
}
\value{
Numeric value of gamma computed as the sum of the column variances
 multiplied by the log of the number of data points
}
\description{
This function can provide a reasonable gamma penalty term for PCF or MultiPCF.
It uses BIC to adjust to the size of the data and the amount of variance in each
sample. It's not magic, though, and might not produce the segmentation you want.

NB: An alternative is to scale the input to unit variance, and then compute
gamma as num_samples * log(num_rows). This has the advantage of making each sample
contribute approximately equally to the segmentation.
}
