% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_API.R
\name{estimate_sd}
\alias{estimate_sd}
\title{Estimate the standard deviation of PCF input data.}
\usage{
estimate_sd(x)
}
\value{
an estimate of SD for each column of the input.
  NB: Assumes that the deviation is constant along the input.
}
\description{
Uses Median Absolute Deviation applied to the first differences
of the data, for robustness to any changes of location present.
}
