% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_private.R
\name{meanCalculator}
\alias{meanCalculator}
\title{Prefix sums for computing means and residuals}
\description{
Prefix sums for computing means and residuals
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-meanCalculator-initialize}{\code{meanCalculator$new()}}
    \item \href{#method-meanCalculator-sum}{\code{meanCalculator$sum()}}
    \item \href{#method-meanCalculator-mean}{\code{meanCalculator$mean()}}
    \item \href{#method-meanCalculator-ssr}{\code{meanCalculator$ssr()}}
    \item \href{#method-meanCalculator-residuals_from_target}{\code{meanCalculator$residuals_from_target()}}
    \item \href{#method-meanCalculator-ssr_rounded_mean}{\code{meanCalculator$ssr_rounded_mean()}}
    \item \href{#method-meanCalculator-clone}{\code{meanCalculator$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanCalculator-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-meanCalculator-initialize}{}}}
\subsection{\code{meanCalculator$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{meanCalculator$new(data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{Numerical vector or matrix}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanCalculator-sum"></a>}}
\if{latex}{\out{\hypertarget{method-meanCalculator-sum}{}}}
\subsection{\code{meanCalculator$sum()}}{
  Compute the sum of the values in the owned data
for segment [start, end]
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{meanCalculator$sum(start, end)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start}}{First index of the data segment}
      \item{\code{end}}{Last index of the data segment}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanCalculator-mean"></a>}}
\if{latex}{\out{\hypertarget{method-meanCalculator-mean}{}}}
\subsection{\code{meanCalculator$mean()}}{
  Compute the mean of the values in the owned data
for segment [start, end]
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{meanCalculator$mean(start, end)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start}}{First index of the data segment}
      \item{\code{end}}{Last index of the data segment}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanCalculator-ssr"></a>}}
\if{latex}{\out{\hypertarget{method-meanCalculator-ssr}{}}}
\subsection{\code{meanCalculator$ssr()}}{
  Compute the sum of squared residuals relative to the mean for
segment [start, end]
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{meanCalculator$ssr(start, end)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start}}{First index of the data segment}
      \item{\code{end}}{Last index of the data segment}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanCalculator-residuals_from_target"></a>}}
\if{latex}{\out{\hypertarget{method-meanCalculator-residuals_from_target}{}}}
\subsection{\code{meanCalculator$residuals_from_target()}}{
  Compute the sum of the squared residuals relative to an arbitrary
target value for segment [start, end]
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{meanCalculator$residuals_from_target(start, end, target)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start}}{First index of the data segment}
      \item{\code{end}}{Last index of the data segment}
      \item{\code{target}}{Residuals are deviations from this target value}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanCalculator-ssr_rounded_mean"></a>}}
\if{latex}{\out{\hypertarget{method-meanCalculator-ssr_rounded_mean}{}}}
\subsection{\code{meanCalculator$ssr_rounded_mean()}}{
  Compute the sum of the squared residuals relative to the integer
rounded mean for segment [start, end]
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{meanCalculator$ssr_rounded_mean(start, end)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start}}{First index of the data segment}
      \item{\code{end}}{Last index of the data segment}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meanCalculator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-meanCalculator-clone}{}}}
\subsection{\code{meanCalculator$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{meanCalculator$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
