% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_API.R
\name{pelt_multipcf}
\alias{pelt_multipcf}
\title{Pruned Exact Linear Time optimisation of Multi PCF}
\usage{
pelt_multipcf(x, kmin, gamma, w = NULL)
}
\arguments{
\item{x}{Matrix of values to segment, samples in columns}

\item{kmin}{Minimum segment size}

\item{gamma}{Penalty term added for each new segment. High values, few segments.}

\item{w}{Optional vector of weights to apply to the columns of x, to up- or down-weight the constribution of each sample to the result.}
}
\value{
List of starts, ends and lengths of segments, and the mean value of x within each.
}
\description{
Pruned Exact Linear Time optimisation of Multi PCF
}
