% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_API.R
\name{pelt_pcf}
\alias{pelt_pcf}
\title{Pruned Exact Linear Time optimisation of PCF}
\usage{
pelt_pcf(x, kmin, gamma)
}
\arguments{
\item{x}{Vector of data to segment}

\item{kmin}{Minimum segment size}

\item{gamma}{Penalty term added for each new segment. Higher values, fewer segments.}
}
\value{
List of starts, ends and lengths of segments, and the mean value of x within each.
}
\description{
Pruned Exact Linear Time optimisation of PCF
}
