% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_API.R
\name{restricted_pcf}
\alias{restricted_pcf}
\title{PCF algorithm, but the algorithm can only place breakpoints at the indices
listed in `breakpoints`}
\usage{
restricted_pcf(x, breakpoints, kmin, gamma)
}
\description{
PCF algorithm, but the algorithm can only place breakpoints at the indices
listed in `breakpoints`
}
