% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_API.R
\name{scale_pcf_input}
\alias{scale_pcf_input}
\title{Scale input data to unit variance.}
\usage{
scale_pcf_input(x)
}
\arguments{
\item{x}{Numeric vector or matrix of PCF inputs}
}
\value{
Data of the same size and shape as the input, scaled so each column
  has unit variance. Vectors are treated as single column matrices.
}
\description{
Scale input data to unit variance.
}
