\name{segmentation-package}
\alias{segmentation-package}
\alias{segmentation}
\docType{package}
\title{
\packageTitle{segmentation}
}
\description{
\packageDescription{segmentation}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{segmentation}
\packageIndices{segmentation}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{segmentation}

Maintainer: \packageMaintainer{segmentation}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
