% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_R_interface.R
\name{multipcf}
\alias{multipcf}
\title{Segment multiple samples simultaneously using the MultiPCF algorithm.}
\usage{
multipcf(x, gamma)
}
\arguments{
\item{x}{Matrix of observations (samples arranged as columns)}

\item{gamma}{Penalty parameter. Higher values produce fewer segments.}
}
\description{
Segment multiple samples simultaneously using the MultiPCF algorithm.
}
