% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_API.R
\name{pcf}
\alias{pcf}
\title{Segment a single sample using the PCF algorithm.}
\usage{
pcf(x, kmin, gamma)
}
\arguments{
\item{x}{Vector of observations}

\item{kmin}{Minimum number of data points permitted in a single segment}

\item{gamma}{Penalty parameter. Higher values produce fewer segments.}
}
\description{
Segment a single sample using the PCF algorithm.
}
